refactor: move split/join tests to dedicated file
Addresses review feedback to reduce the size of
positronNotebookInstance.test.ts by extracting all split and join cell
tests into positronNotebookSplitJoin.test.ts.
